Added feature "Check in".
This feature help TA and teacher to know who visit the student calling.

+++ src/main/java/org/mklab/taskit/server/domain/AbstractEntity.java
@@ -108,7 +108,9 @@ public abstract class AbstractEntity<I> {
     try {
       t.begin();
       final AbstractEntity<?> entity = em.find(getClass(), getId());
-      em.remove(entity);
+      if (entity != null) {
+        em.remove(entity);
+      }
       t.commit();
     } catch (Throwable ex) {
       ex.printStackTrace();
